+++ contracts/OwnableOZ.sol
@@ -8,6 +8,7 @@ pragma solidity ^0.4.21;
 contract OwnableOZ
 {
 	address public m_owner;
+	bool    public changeable;
 
 	event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
 
@@ -27,17 +28,20 @@ contract OwnableOZ
 	function OwnableOZ() public
 	{
 		m_owner = msg.sender;
+		changeable =true;
 	}
 
 	/**
 	 * @dev Allows the current owner to transfer control of the contract to a newOwner.
 	 * @param _newOwner The address to transfer ownership to.
 	 */
-	function transferOwnership(address _newOwner) public onlyOwner
+	function setImmutableOwnership(address _newOwner) public onlyOwner
 	{
+		require(changeable);
 		require(_newOwner != address(0));
-		emit OwnershipTransferred(m_owner, _newOwner);
 		m_owner = _newOwner;
+		changeable = false;
+		emit OwnershipTransferred(m_owner, _newOwner);
 	}
 
 }